Fixed two invalid memory accesses. Also fixed the audio.
One caused from accessing the same memory twice in the same function
with imgui (not sure root cause).

Second one was caused by assigning an entity a component to the wrong
eid, when constructing water instances.

These two fixes inadvertenly fixed the broken audio bug with the water
not playing sound when the player was in the water.