Working towards spawning entities after level loading time.
Specifically working towards being able to concate to the list of
entities being drawn to the screen per-frame. Presently it only is
modified during level assembling.

Part of this work is splitting up the gpu/factory inside the opengl/
module, so it's clearer when the vertices are being created / copied to
the GPU. Right now those two things are mixed into both the factory.hpp
and gpu.hpp.