Fixed the debug crash inside waterinfo binding.
There was a logic error with the debug binding macros. If a debug object was
bound, then moved from, sometimes depending on the order of things at
runtime, the scoped macros would try unbinding the now-moved-from debug
object (who's move constructor turned is_bound to false) causing a crash
when the macro went to re-unbind the now-moved-from debug object.

Solution is to add a third state, MOVED-FROM, and check it during
binding. If an object was moved-from, it is assumed the new object will
take over the responsibility of unbinding it.