fixed the invalid memory access I've been chasing
Root cause: The memory where the DrawInfo's were being stored (for the
WaterInfo's) was being resized after the WaterInfo's dinfo pointer was
set.

This wasn't a problem for things like Terrain which, which own's it's
DrawInfo. Presently, WaterInfo instances are stored on the registry so
they cannot own a DrawInfo.