Fixed bug preventing us from moving the "StaticRenders".
Found a bug while trying to make the StaticRenders intance not-static.
I was binding the framebuffer's incorrectly, but since I called Imgui's
NewFrame() function every frame, it was fixing this bug.

When I moved the instance of StaticRenderers to before the call to
Imgui::NewFrame(), I was getting an OpenGL state error. Fixing the
framebuffer binding bug allows the StaticRenderers instance to be put
onto the stack.